+++ Runtime/Scripts/Components/InstantiateOnceOnRuntime.cs
@@ -1,0 +1,31 @@
+using System;
+using Unity.Scripting.LifecycleManagement;
+using UnityEngine;
+
+namespace ParkMinPackages.Workflow.Default.Components
+{
+	[DefaultExecutionOrder(int.MinValue)]
+	public partial class InstantiateOnceOnRuntime : MonoBehaviour
+	{
+		[AutoStaticsCleanup]
+		static bool IsInstantiated = false;
+
+		void Awake() {
+			if (IsInstantiated)
+				return;
+
+			foreach (GameObject prefab in _prefabs) {
+				try {
+					Instantiate(prefab);
+				}
+				catch (Exception e) {
+					Debug.LogError(e);
+				}
+			}
+			IsInstantiated = true;
+		}
+
+
+		[SerializeField] GameObject[] _prefabs;
+	}
+}