+++ Runtime/Scripts/Objects/PrefabKindGroups.cs
@@ -1,0 +1,9 @@
+using Sirenix.OdinInspector;
+
+namespace ParkMinPackages.Workflow.Default.Objects
+{
+	public static class PrefabKindGroups
+	{
+		public const PrefabKind SceneObject = PrefabKind.InstanceInScene | PrefabKind.NonPrefabInstance;
+	}
+}